--- a/lib/power_trace/entries/block_entry.rb
+++ b/lib/power_trace/entries/block_entry.rb
@@ -1,4 +1,0 @@
-module PowerTrace
-  class BlockEntry < Entry
-  end
-end