--- a/lib/power_trace/entry.rb
+++ b/lib/power_trace/entry.rb
@@ -132,8 +132,23 @@ module PowerTrace
 
       [locals, arguments]
     end
+
+    def method
+      @method ||= Object.instance_method(:method).bind(@receiver).call(name)
+    rescue NameError
+      # if any part of the program uses Refinement to extend its methods
+      # we might still get NoMethodError when trying to get that method outside the scope
+      nil
+    end
+
+    private
+
+    def method_parameters
+      if method
+        method.parameters.map { |parameter| parameter[1] }
+      else
+        []
+      end
+    end
   end
 end
-
-require "power_trace/entries/method_entry"
-require "power_trace/entries/block_entry"