… lib/power_trace/entries/method_entry.rb