### Performance Boost!

It turned out that the major performance bottleneck is the case statement for `Entry` initialization. After removing that, the performance improved by 3~6x when processing higher number of traces!

#### Before

```
  1000 times - ruby 2.7.1
  Normal raise
     0.001606   0.000091   0.001697 (  0.001692)
     0.000953   0.000082   0.001035 (  0.001035)
     0.000885   0.000099   0.000984 (  0.000984)
  ====================================
  ======= power_trace required =======
  ====================================
  Raise and store power_trace (trace_limit: 10)
     0.197135   0.000762   0.197897 (  0.198118)
     0.203479   0.001198   0.204677 (  0.205063)
     0.197096   0.000967   0.198063 (  0.198181)
  Raise and store power_trace (trace_limit: 50)
     1.187342   0.003542   1.190884 (  1.191932)
     1.183849   0.003203   1.187052 (  1.187759)
     1.196402   0.002802   1.199204 (  1.199782)
  Raise and replace backtrace (trace_limit: 10)
     0.406225   0.001383   0.407608 (  0.408063)
     0.405999   0.000521   0.406520 (  0.406738)
     0.411041   0.001345   0.412386 (  0.412863)
  Raise and replace backtrace (trace_limit: 50)
     1.528390   0.004480   1.532870 (  1.534161)
     1.544761   0.003519   1.548280 (  1.549895)
     1.387785   0.004530   1.392315 (  1.393240)
```

#### After
```
1000 times - ruby 2.7.1
Normal raise
   0.002032   0.000093   0.002125 (  0.002119)
   0.001131   0.000107   0.001238 (  0.001239)
   0.001092   0.000100   0.001192 (  0.001192)
====================================
======= power_trace required =======
====================================
Raise and store power_trace (trace_limit: 10)
   0.157296   0.001807   0.159103 (  0.159504)
   0.159911   0.000668   0.160579 (  0.160961)
   0.152814   0.000493   0.153307 (  0.153658)
Raise and store power_trace (trace_limit: 50)
   0.228311   0.000649   0.228960 (  0.229264)
   0.248271   0.001956   0.250227 (  0.257272)
   0.229469   0.000923   0.230392 (  0.230822)
Raise and replace backtrace (trace_limit: 10)
   0.331577   0.000360   0.331937 (  0.332203)
   0.354934   0.001634   0.356568 (  0.356939)
   0.334688   0.001201   0.335889 (  0.336682)
Raise and replace backtrace (trace_limit: 50)
   0.557920   0.002530   0.560450 (  0.561101)
   0.553058   0.001767   0.554825 (  0.555651)
   0.560771   0.002160   0.562931 (  0.563613)
```

### PowerTrace.trace_limit
This PR also adds the `PowerTrace.trace_limit` for limiting trace computation. But as you can see from the above result, it doesn't affect the total computation time in a linear way. And the default number `30` shouldn't affect most use cases, it's just a safeguard to make sure we don't spend too much time on power trace collection, even in rare edge cases.